Refactor: break down main.rs into manageable modules
Extract logical components from main.rs into separate modules for
improved readability and maintainability:

- cli.rs: CLI argument parsing (Args, DaemonModeFlag)
- health.rs: Health check HTTP server and handlers
- daemon.rs: Daemon mode logic and event loop
- svid.rs: Shared X.509 certificate fetching logic (DRY)
- lib.rs: Module exports for library crate

The binary (main.rs) now imports from the library crate instead of
declaring modules separately, avoiding dual compilation.

Also moves workload_api tests to tests/ directory for better
separation of unit and integration tests.

Changes:
- src/main.rs: 313 -> 46 lines (orchestration only, uses library)
- src/cli.rs: New file (35 lines)
- src/health.rs: New file (52 lines)
- src/daemon.rs: New file (82 lines)
- src/svid.rs: New file (shared fetch logic + tests)
- src/lib.rs: New file (6 lines)
- tests/workload_api_tests.rs: Moved from src/workload_api.rs
- Cargo.toml: Added lib target configuration
