+++ Cargo.toml
@@ -3,6 +3,14 @@ name = "spiffe-helper-rust"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "spiffe_helper_rust"
+path = "src/lib.rs"
+
+[[bin]]
+name = "spiffe-helper-rust"
+path = "src/main.rs"
+
 [dependencies]
 clap = { version = "4.5", features = ["derive"] }
 anyhow = "1.0"
