+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod cli;
+pub mod config;
+pub mod daemon;
+pub mod health;
+pub mod svid;
+pub mod workload_api;
