+++ src/main.rs
@@ -1,44 +1,14 @@
-mod config;
-mod workload_api;
-
 use anyhow::{Context, Result};
-use axum::{http::StatusCode, response::IntoResponse, routing::get, Router};
-use clap::{Parser, ValueEnum};
+use clap::Parser;
 use std::path::PathBuf;
-use tokio::signal::unix::{signal, SignalKind};
-use tokio::time::{interval, Duration};
+
+use spiffe_helper_rust::{cli, config, daemon, svid};
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
-const DEFAULT_CONFIG_FILE: &str = "helper.conf";
-const DEFAULT_LIVENESS_LOG_INTERVAL_SECS: u64 = 30;
-
-#[derive(Copy, Clone, Debug, ValueEnum)]
-enum DaemonModeFlag {
-    True,
-    False,
-}
-
-/// SPIFFE Helper - A utility for fetching X.509 SVID certificates from the SPIFFE Workload API
-#[derive(Parser, Debug)]
-#[command(name = "spiffe-helper-rust")]
-#[command(about = "SPIFFE Helper - Fetch and manage X.509 SVID certificates", long_about = None)]
-struct Args {
-    /// Path to the configuration file
-    #[arg(short, long, default_value = DEFAULT_CONFIG_FILE)]
-    config: String,
-
-    /// Boolean true or false. Overrides `daemon_mode` in the config file.
-    #[arg(long, value_enum)]
-    daemon_mode: Option<DaemonModeFlag>,
-
-    /// Print version number
-    #[arg(short = 'v', long)]
-    version: bool,
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let args = Args::parse();
+    let args = cli::Args::parse();
 
     // Handle version flag
     if args.version {
@@ -47,10 +17,7 @@ async fn main() -> Result<()> {
     }
 
     // Parse daemon_mode override
-    let daemon_mode_override = args.daemon_mode.map(|flag| match flag {
-        DaemonModeFlag::True => true,
-        DaemonModeFlag::False => false,
-    });
+    let daemon_mode_override = args.daemon_mode.map(cli::DaemonModeFlag::to_bool);
 
     // Parse config file
     let config_path = PathBuf::from(&args.config);
@@ -65,249 +32,15 @@ async fn main() -> Result<()> {
     // Check if daemon mode is enabled (defaults to true)
     let daemon_mode = config.daemon_mode.unwrap_or(true);
     if daemon_mode {
-        // Run daemon mode
-        run_daemon(config).await
+        daemon::run(config).await
     } else {
-        // Non-daemon mode - fetch certificates once and exit
         run_once(config).await
     }
 }
 
 async fn run_once(config: config::Config) -> Result<()> {
     println!("Running spiffe-helper-rust in one-shot mode...");
-
-    fetch_x509_certificate(&config).await?;
+    svid::fetch_x509_certificate(&config).await?;
     println!("One-shot mode complete");
     Ok(())
-}
-
-/// Fetches the initial X.509 SVID from the SPIRE agent and writes it to the configured directory.
-///
-/// This function validates the configuration (`agent_address` and `cert_dir`) and calls
-/// `workload_api::fetch_and_write_x509_svid` to perform the actual fetch and write operation.
-/// It implements the shared initial SVID fetch policy used by both daemon and one-shot modes,
-/// including retry logic and backoff handling.
-///
-/// # Arguments
-///
-/// * `config` - The configuration containing agent address, cert directory, and file names
-///
-/// # Returns
-///
-/// Returns `Ok(())` if successful, or an error if configuration is invalid or fetching fails.
-async fn fetch_x509_certificate(config: &config::Config) -> Result<()> {
-    let agent_address = config
-        .agent_address
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("agent_address must be configured"))?;
-    let cert_dir = config
-        .cert_dir
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("cert_dir must be configured"))?;
-
-    println!("Fetching X.509 certificate from SPIRE agent at {agent_address}...");
-    let cert_dir_path = std::path::PathBuf::from(cert_dir);
-    workload_api::fetch_and_write_x509_svid(
-        agent_address,
-        &cert_dir_path,
-        config.svid_file_name.as_deref(),
-        config.svid_key_file_name.as_deref(),
-    )
-    .await
-    .with_context(|| "Failed to fetch X.509 certificate")?;
-    println!("Successfully fetched and wrote X.509 certificate to {cert_dir}");
-    Ok(())
-}
-
-async fn run_daemon(config: config::Config) -> Result<()> {
-    println!("Starting spiffe-helper-rust daemon...");
-
-    // Fetch initial X.509 SVID at startup
-    // This ensures certificates are available before the daemon continues
-    fetch_x509_certificate(&config).await?;
-
-    // Start health check server if enabled
-    let health_checks = config.health_checks.clone();
-    let mut health_server_handle = if let Some(ref health_checks) = health_checks {
-        if health_checks.listener_enabled {
-            let bind_addr = format!("0.0.0.0:{}", health_checks.bind_port);
-            let liveness_path = health_checks
-                .liveness_path
-                .clone()
-                .unwrap_or_else(|| "/health/live".to_string());
-            let readiness_path = health_checks
-                .readiness_path
-                .clone()
-                .unwrap_or_else(|| "/health/ready".to_string());
-
-            println!("Starting health check server on {bind_addr}");
-            println!("  Liveness path: {liveness_path}");
-            println!("  Readiness path: {readiness_path}");
-
-            let app = Router::new()
-                .route(&liveness_path, get(liveness_handler))
-                .route(&readiness_path, get(readiness_handler));
-
-            let listener = tokio::net::TcpListener::bind(&bind_addr)
-                .await
-                .with_context(|| format!("Failed to bind to {bind_addr}"))?;
-
-            Some(tokio::spawn(async move {
-                axum::serve(listener, app)
-                    .await
-                    .context("Health check server failed")
-            }))
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    // Set up signal handling for graceful shutdown
-    let mut sigterm =
-        signal(SignalKind::terminate()).context("Failed to register SIGTERM handler")?;
-
-    // Set up periodic liveness logging
-    let mut liveness_interval = interval(Duration::from_secs(DEFAULT_LIVENESS_LOG_INTERVAL_SECS));
-    liveness_interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-
-    println!("Daemon running. Waiting for SIGTERM to shutdown...");
-
-    // Main daemon loop
-    let result = loop {
-        tokio::select! {
-            _ = sigterm.recv() => {
-                println!("Received SIGTERM, shutting down gracefully...");
-                break Ok(());
-            }
-            _ = liveness_interval.tick() => {
-                println!("spiffe-helper-rust daemon is alive");
-            }
-            // If health server is running, watch for its completion (which indicates failure)
-            Some(res) = async {
-                match &mut health_server_handle {
-                    Some(handle) => Some(handle.await),
-                    None => std::future::pending().await,
-                }
-            } => {
-                match res {
-                    Ok(Ok(())) => {
-                        // Server exited cleanly (shouldn't happen normally)
-                        println!("Health check server exited unexpectedly");
-                        break Ok(());
-                    }
-                    Ok(Err(e)) => {
-                        // Server returned an error
-                        break Err(e);
-                    }
-                    Err(e) => {
-                        // Task panicked or was cancelled
-                        break Err(anyhow::anyhow!("Health check server task failed: {e}"));
-                    }
-                }
-            }
-        }
-    };
-
-    // Shutdown health check server if it was started and still running
-    if let Some(ref handle) = health_server_handle {
-        if !handle.is_finished() {
-            handle.abort();
-            println!("Health check server stopped");
-        }
-    }
-
-    println!("Daemon shutdown complete");
-    result
-}
-
-async fn liveness_handler() -> impl IntoResponse {
-    StatusCode::OK
-}
-
-async fn readiness_handler() -> impl IntoResponse {
-    StatusCode::OK
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_fetch_x509_certificate_missing_agent_address() {
-        let config = config::Config {
-            agent_address: None,
-            cert_dir: Some("/tmp/certs".to_string()),
-            svid_file_name: None,
-            svid_key_file_name: None,
-            ..Default::default()
-        };
-
-        let result = fetch_x509_certificate(&config).await;
-        assert!(result.is_err());
-        let error_msg = result.unwrap_err().to_string();
-        assert!(error_msg.contains("agent_address must be configured"));
-    }
-
-    #[tokio::test]
-    async fn test_fetch_x509_certificate_missing_cert_dir() {
-        let config = config::Config {
-            agent_address: Some("unix:///tmp/agent.sock".to_string()),
-            cert_dir: None,
-            svid_file_name: None,
-            svid_key_file_name: None,
-            ..Default::default()
-        };
-
-        let result = fetch_x509_certificate(&config).await;
-        assert!(result.is_err());
-        let error_msg = result.unwrap_err().to_string();
-        assert!(error_msg.contains("cert_dir must be configured"));
-    }
-
-    #[tokio::test]
-    async fn test_fetch_x509_certificate_with_custom_file_names() {
-        // This test verifies that custom file names are passed through correctly
-        // We can't easily test the full flow without a SPIRE agent, but we can
-        // verify the configuration validation works with custom file names
-        let config = config::Config {
-            agent_address: Some("unix:///tmp/nonexistent-agent.sock".to_string()),
-            cert_dir: Some("/tmp/certs".to_string()),
-            svid_file_name: Some("custom_cert.pem".to_string()),
-            svid_key_file_name: Some("custom_key.pem".to_string()),
-            ..Default::default()
-        };
-
-        // This will fail when trying to connect to the agent, but that's expected
-        // The important part is that it validates config correctly first
-        let result = fetch_x509_certificate(&config).await;
-        assert!(result.is_err());
-        // Should fail on connection, not on config validation
-        let error_msg = result.unwrap_err().to_string();
-        // Should not be a config validation error
-        assert!(!error_msg.contains("agent_address must be configured"));
-        assert!(!error_msg.contains("cert_dir must be configured"));
-    }
-
-    #[test]
-    fn test_greeting_logic() {
-        // Mock test to ensure coverage data is generated
-        // Test the greeting logic used in main()
-        let name = Some("Rust".to_string());
-        let greeting_name = name.as_deref().unwrap_or("World");
-        assert_eq!(greeting_name, "Rust");
-
-        let no_name: Option<String> = None;
-        let default_name = no_name.as_deref().unwrap_or("World");
-        assert_eq!(default_name, "World");
-    }
-
-    #[test]
-    fn test_coverage_helper() {
-        // Additional test to generate coverage data
-        let count: u8 = 3;
-        assert!(count > 0);
-        assert_eq!(count, 3);
-    }
 }
